Generate a sequence of G values

Generate a sequence of G values based on the specified parameters.

Usage

generate_sequence(phi, lambda, eta)

Arguments

phi

A numeric value representing the correlation parameter between two contious observations.

lambda

A numeric value representing the place where chang point happens.

eta

A numeric value representing the constants when generating 'yt' based on 'xt'.

Value

A vector of G values.

Examples

generate_sequence(phi = 0.5, lambda = 0.2, eta = 1)
#>  [1] 20.074508 13.726827 31.607955  2.591526  2.818693  8.596335  2.757768
#>  [8] 88.372764 49.638236  6.307562  2.395732  3.248580  8.831463  9.093505
#> [15]  5.974588 15.731831  2.808938 14.867489  6.325023  2.617496  4.702790
#> [22]  3.849700 14.187629  2.624147 15.013495  7.448054 49.896203 18.811895
#> [29]  8.559212 10.377157  4.922793  2.559862  3.106018 28.070552 33.588857
#> [36]  4.968885 12.382653 15.640018  3.313328 11.002739  8.381014 14.416079
#> [43] 13.728812  2.258115  3.241911 20.554219 11.715762  5.956822  6.001568
#> [50]  3.180278